Fixed some bugs and added the "-#" flag to curl
I promised myself I'd be a good programmer and keep the master branch
more stable, so off this code goes into the testing branch. I also added
a --with-pid flag option to --all which prepends the pid number before
the command used with it as well as some other code changes/fixes.